# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bus Booking System",
        "version": "0.1"
    },
    "paths": {
...
        "/api/trips": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "description": "retrieve Filtered Trips",
                "operationId": "34b92215d6018d9559a016ee14964292",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "trip date ",
                        "required": true,
                        "example": "2023-06-02"
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "source",
                        "required": true,
                        "example": 1
                    },
                    {
                        "name": "destination",
                        "in": "query",
                        "description": "destination",
                        "required": true,
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "If any error happened",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Not Authenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/api/trips/{trip_id}": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "description": "show trip details with available seats",
                "operationId": "175cc7566c98b78a46ee9fbe7cff7600",
                "parameters": [
                    {
                        "name": "trip_id",
                        "in": "path",
                        "description": "trip id ",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "If any error happened",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Not Authenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/api/trips/available-seats": {
            "post": {
                "tags": [
                    "Trips"
                ],
                "description": "get Available Seats",
                "operationId": "7c56eed96538d8bd41c7d0fa5f08a1d3",
                "requestBody": {
                    "description": "trip search paramters (source , destination , date )",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "source",
                                    "destination",
                                    "date"
                                ],
                                "properties": {
                                    "trip_id": {
                                        "type": "interger",
                                        "format": "text",
                                        "example": "1"
                                    },
                                    "source_station_id": {
                                        "type": "interger",
                                        "format": "text",
                                        "example": "1"
                                    },
                                    "destination_station_id": {
                                        "type": "interger",
                                        "format": "text",
                                        "example": "3"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-06-02"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Available Seats",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Success",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Available Seats"
                                        },
                                        "data": {
                                            "description": "Array of Trips and available seats inside",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "format": "text",
                                            "example": "You Are Not Authorized!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Fill Data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "jwt": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}